+++ src/Handlers/Logging/Observers/Monolog.php
@@ -5,9 +5,12 @@ namespace Buckaroo\Handlers\Logging\Observers;
 use Buckaroo\Handlers\Logging\Observer;
 use Monolog\Handler\StreamHandler;
 use Monolog\Logger;
+use Psr\Log\LoggerInterface;
 
 class Monolog implements Observer
 {
+    protected LoggerInterface $log;
+
     public function __construct()
     {
         $this->log = new Logger('Buckaroo log');