+++ src/Handlers/Reply/HttpPost.php
@@ -51,8 +51,7 @@ class HttpPost implements ReplyStrategy
     public function validate(): bool
     {
         //Remove brq_signature from the equation
-        $data = array_filter($this->data, function ($key)
-        {
+        $data = array_filter($this->data, function ($key) {
             $acceptable_top_level = ['brq', 'add', 'cust', 'BRQ', 'ADD', 'CUST'];
 
             return (
@@ -61,8 +60,7 @@ class HttpPost implements ReplyStrategy
         }, ARRAY_FILTER_USE_KEY);
 
         //Combine the array keys with value
-        $data = array_map(function ($value, $key)
-        {
+        $data = array_map(function ($value, $key) {
             return $key . '=' . html_entity_decode($value);
         }, $data, array_keys($data));