+++ src/PaymentMethods/PaymentMethod.php
@@ -28,6 +28,9 @@ use Buckaroo\Transaction\Request\TransactionRequest;
 use Buckaroo\Transaction\Response\TransactionResponse;
 use Psr\Log\LoggerInterface;
 
+/**
+ *
+ */
 abstract class PaymentMethod implements PaymentInterface
 {
     /**
@@ -73,6 +76,11 @@ abstract class PaymentMethod implements PaymentInterface
      * @var bool
      */
     protected bool $isManually = false;
+
+    /**
+     * @var string|null
+     */
+    protected ?string $serviceCode;
 
     /**
      * @param Client $client
@@ -185,8 +193,7 @@ abstract class PaymentMethod implements PaymentInterface
     {
         $this->combinablePayment = $combinablePayment;
 
-        $payload_data = array_filter($combinablePayment->request->data(), function ($key)
-        {
+        $payload_data = array_filter($combinablePayment->request->data(), function ($key) {
             return ! in_array($key, ['Services']);
         }, ARRAY_FILTER_USE_KEY);