+++ src/Transaction/Request/HttpClient/HttpClientGuzzle.php
@@ -25,6 +25,7 @@ namespace Buckaroo\Transaction\Request\HttpClient;
 use Buckaroo\Exceptions\TransferException;
 use Buckaroo\Handlers\Logging\Subject;
 use GuzzleHttp\Client;
+use GuzzleHttp\ClientInterface;
 use GuzzleHttp\Exception\GuzzleException;
 use GuzzleHttp\Psr7\Request;
 use GuzzleHttp\RequestOptions;
@@ -35,6 +36,8 @@ class HttpClientGuzzle extends HttpClientAbstract
      * @var Subject
      */
     protected Subject $logger;
+
+    protected ClientInterface $httpClient;
 
     /**
      * @param Subject $logger
@@ -66,17 +69,14 @@ class HttpClientGuzzle extends HttpClientAbstract
 
         $request = new Request($method, $url, $headers, $data);
 
-        try
-        {
+        try {
             $response = $this->httpClient->send($request, ['http_errors' => false]);
 
             $result = (string) $response->getBody();
 
             $this->logger->info('RESPONSE HEADERS: ' . json_encode($response->getHeaders()));
             $this->logger->info('RESPONSE BODY: ' . $response->getBody());
-        }
-        catch (GuzzleException $e)
-        {
+        } catch (GuzzleException $e) {
             throw new TransferException($this->logger, "Transfer failed", 0, $e);
         }